Split response to base response and response, will use the base response to define streaming response
